requirements.txt
fastapi>=0.135,<1.0
uvicorn[standard]>=0.44,<1.0
sqlalchemy>=2.0,<3.0
pydantic>=2.13,<3.0
python-multipart>=0.0.26,<1.0
qdrant-client>=1.17,<2.0
onnxruntime>=1.24,<2.0
pillow>=12.2,<13.0
numpy>=2.4,<3.0
opencv-python-headless>=4.10,<5.0
